+++ projects/backend/Beta/ProvaUfficiale/src/main/java/interfaces/UtenteAppInterface.java
@@ -1,5 +1,6 @@
 package interfaces;
 
+import java.util.Date;
 import java.util.List;
 
 import model.Evento;
@@ -9,5 +10,8 @@ public interface UtenteAppInterface {
 	
 	abstract List<Evento> visualizzaEventiUtente(UtenteApp utente);
 	abstract List<Evento> visualizzaTuttiEventi();
+
+	public abstract void nuovoEvento( String nomeEvento,  Date dataDiNascita, Date dataDiRegistrazione, String descrizione, String latitudine, String longitudine);
+	
 
 }
